requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.132
botocore==1.34.132
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-prometheus==2.3.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
exceptiongroup==1.2.1
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.7
olefile==0.47
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytest==8.2.2
pytest-django==4.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
